Revert "Solutions/13 integration testing"
