+++ cypress/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  root: true,
+  plugins: ['eslint-plugin-cypress'],
+  extends: ['react-app', 'plugin:cypress/recommended'],
+  env: {'cypress/globals': true},
+}
