+++ cypress/support/e2e.js
@@ -1,0 +1,2 @@
+import 'cypress-hmr-restarter'
+import '@testing-library/cypress/add-commands'
