+++ cypress/support/generate.js
@@ -1,0 +1,1 @@
+export * from '../../src/test/generate'
