+++ netlify.toml
@@ -1,3 +1,5 @@
 [build]
   command = "npm run validate"
   publish = "build"
+[[plugins]]
+  package = "netlify-plugin-cypress"
