+++ src/utils/__tests__/api-client.js
@@ -6,7 +6,7 @@ import {client} from '../api-client'
 const apiURL = process.env.REACT_APP_API_URL
 
 jest.mock('react-query')
-jest.mock('auth-provider', () => jest.createMockFromModule('auth-provider'))
+jest.mock('auth-provider')
 
 test('calls fetch at the endpoint with the arguments for GET requests', async () => {
   const endpoint = 'test-endpoint'
